메일 엔티티,dto 수정2차

+++ src/main/java/com/example/companycore/service/MessageApiClient.java
@@ -181,7 +181,7 @@ public class MessageApiClient extends BaseApiClient {
      * @param userId 사용자 ID (헤더에 포함)
      * @return 메시지 DTO 객체
      */
-    public MessageDto getMessageById(Long messageId, Long userId) {
+    public MessageDto getMessageById(Integer messageId, Long userId) {
         try {
             HttpRequest request = createAuthenticatedRequestBuilder("/messages/" + messageId)
                     .header("User-Id", userId.toString())
@@ -207,7 +207,7 @@ public class MessageApiClient extends BaseApiClient {
      * ✅ 메시지 상태 업데이트 (PUT /messages/{id}).
      * 예: 읽음 처리, 삭제 등
      */
-    public boolean updateMessageStatus(Long messageId, Long userId, String action) {
+    public boolean updateMessageStatus(Integer messageId, Long userId, String action) {
         try {
             ObjectNode body = objectMapper.createObjectNode();
             body.put("action", action); // 예: "read", "delete"
@@ -255,7 +255,7 @@ public class MessageApiClient extends BaseApiClient {
     /**
      * ✅ 특정 메시지에 답장합니다 (POST /messages/{id}/reply).
      */
-    public MessageDto replyToMessage(Long messageId, Long userId, String title, String content) {
+    public MessageDto replyToMessage(Integer messageId, Long userId, String title, String content) {
         try {
             ObjectNode body = objectMapper.createObjectNode();
             body.put("title", title);
@@ -354,7 +354,7 @@ public class MessageApiClient extends BaseApiClient {
      * @param userId 사용자 ID (헤더로 전달)
      * @return 삭제 성공 여부
      */
-    public boolean deleteMessageById(Long messageId, Long userId) {
+    public boolean deleteMessageById(Integer messageId, Long userId) {
         try {
             HttpRequest request = createAuthenticatedRequestBuilder("/messages/" + messageId)
                     .header("User-Id", userId.toString())
